♻️ Refactor: 키보드 내리기 확장 함수화
#Related to: #252
